Change min reviews required for RXP calculation
Every current non-SEP player will be given an external review count of 1 to start, instead of 0, because it's easier to do that than to teach the code to handle the special case of a player somehow having earned some accuracy without ever submitting a review. 

Since that moves everyone one review closer to the minimum needed to calculate RXP, I've changed that threshold from 7 to 8.